Switch spidev to specific device in devicetree
Originally the device tree entries for each spidev was specified as
'spidev' in the compatible string. However, while this originally just
warned, it is now ignored. The kernel insists that the compatible string
actually specifies a part instead.

Since we're not actually controlling a specific part, we must just
include any of the parts recognised as a spidev (see comment in device
tree). In this case a Rohm device was chosen, though the actual device
is irrelevant. The /dev/spidev*.* devices now appear again.