Add udev rules for I2C named buses
Each I2C bus on the muxes now has a symlinked version in /dev/. This
makes it far easier to find the right one. You can still use the old
numbered ones if you wish.

This is compatible with odin-devices I2CDevice, since the smbus2 library
underneath supports receiving the bus as a path as well as a number.

This mechanism of using an actual rules file directly for udev has also
worked well. It is clearer and easier to manage, and simplifies the
recipe. I will replace the other manually created ones shortly.