Change udev rule installation to use separate files
The original mechanism for this was clumsy use of 'echo', which made
things confusing and more annoying to update.
This way, any udev rules file included in the new rules directory will
be included in the build.

The execution time of the rules for I2C have also been updated so that
they run before the odin instances; this caused issues with access that
then covered up the evidence by changing ownership before I was able to
check what was going on...

Also updated loki-config to use a directory include, which is far
cleaner.