Fix flash mounting issues
QSPI flash is now detected as the correct size, and has an additional partition large enough to be mounted for general user use since the erase block size has changed due to kernel changes leading to an unmountable mtd1. We now use mtd4, which is larger and created specifrically for user use.