Switch network configuration to wired.network and fix systemd unit file dependencies
Previously dependencies for the unit files were quite muddled and some
were unnecessary. This now locks them down more sensibly, with only
those that depend on the network doing so, and those that depend on
filesystems doing so.

The networking has changed from interfaces to wired.network (systemd
networkd), and so the handling of static IP aliases etc has changed.

In addition, the network override that takes place in the eMMC bootstrap
is now forced to take place after the eMMC is mounted, but before the
network init takes place.