Overhaul app and loki version gathering using fdt in Makefile
Originally, the application and loki version information was provided as
a variable at Autoconf configure time, and then built into the image via
a device tree dtsi file, which was generated specifically for this
purpose.

This was prone to error because it meant that the git describe based
version numbers were only updated if the entire project was
re-configured, which was annoying and meant that often there were wildly
different actual compiled versions of the software compared to the
version they reported; it was only really useful to report versions in
images completely cleaned and rebuilt for the official 'releases'.

This process has now been moved to direct modification of the system dtb
using `fdtput` in the makefile. This means that it is exactly up to date
at the time of packaging for both the application and core repo.

In addition, the timestamp stored in the top-level FIT image (image.ub)
is now overridden with the packaging time from the host system. This was
necessary because the one automatically compiled in in the 2023
toolchain appears not to work properly.

For the sake of keeping everything in the same place and simplifying the
design, the other configure-time information (application name, platform
information etc) is also now built in from the Makefile. This means the
metadata dtsi file could be removed completely.

This requires a very small change in the application Makefile: it must
make sure that APPLICATION_DIR is defined (this is already usually the
case) but it must also have an `export` line so that it propagates down
and can be used for the describe. If this is not provided or the
application does not yet have a valid description, the field will be
left blank.