Relocated loki metadata to the top level FIT
The original method of editing the system.dtb worked, but only for
editing the tree used by U-Boot. Instead, the modifications are now to
the image.ub directly (now inline with the timestamp). This means that
the new images will require an update to loki-update to properly read
the new location. This will be backwards compatible, but old images will
not be able to read the version of newer images.

This change should ideally be included in the older toolchain unless the
2023 toolchain is quickly merged for other projects.