Increase boot partition to 8M to fit larger bootloader
The bootloader has increased to just over 4M (the original partition
size) in the new toolchain. This means it will no longer fit when
programmed into the flash.

The most future-proofed method to fix this is to increase, comfortably,
the size of the partition with a lot of headroom for the future. I've
therefore doubled it.

Scripts have been updated, but only for tweaks and unrelated
optimisation; they should otherwise work because the partition numbering
is still all the same. This does however mean that images from this
point forwards will not be able to detect the presence of images
installed into the flash for older versions, and vice versa.