Moved layer config for included image parts to layer.conf
Previously, these were stored in petalinuxbsp.conf, but I think there is
an issue where this is unpredictably combined between layers. The
layer.conf (which is from Yocto, not PetaLinux) is specifically there to
exist in multiple layers and be processed according to the hierarchy.

This seems to have sorted issues where loki-update and other parts of
the image were sometimes not being included.