Fix flash boot offset to match new partitions
Since the flash partition size for the bootloader increased, the
location of the kernel changed. This is now correct.

Note that the method for doing this has changed drastically; since
2023.2, the PetaLinux tools will override the offset and size variables
used in the boot scripts, meaning changes in the Yocto recipes are
ignored.

This is nicer as it can be configured in the menuconfig, but is
currently a little confusing due to variables appearing in multiple
places. My next step is to remove anything unnecessary from the recipe
layer, but I will commit the working system first.