Build odin-sequencer-rpc without pydantic support
Issues with pydantic>2 support have been avoided by removing the
requirement from odin-sequencer. This now installs using poetry (a must
since setuptools3 in this version does not support pyproject.toml, and
there is no setup.py in this project.

The RPC recipe is included automatically with RRECOMMENDS, which can be
overridden if desired.