Prevent root ssh login by disabling debug-tweaks
No matter what was entered into the dropbear.default environment file,
it was being overridden by the image features setting 'debug-tweaks',
which is designed specifically to allow easier and less secure SSH
connections to the device. It will seek out the option to disable root
ssh connections and actually remove it.

This has now been disabled, so the file is left alone, and root
connections are not allowed over SSH. This matches the former setup in
the 2020 toolchain.